Split ConcurrentQueue<T>.Segment into its own file
