+++ src/System.Collections.Concurrent/src/System.Collections.Concurrent.csproj
@@ -22,6 +22,9 @@
     <Compile Include="System\Collections\Concurrent\Partitioner.cs" />
     <Compile Include="System\Collections\Concurrent\PartitionerStatic.cs" />
     <Compile Include="System\Collections\Concurrent\PlatformHelper.cs" />
+    <Compile Include="$(CommonPath)\System\Collections\Concurrent\ConcurrentQueue_Segment.cs">
+      <Link>Common\System\Collections\Concurrent\ConcurrentQueue_Segment.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <Reference Include="System.Collections" />
